Introducing Security namespace - refactoring
