+++ Source/Tenancy.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using Aksio.Cratis.Execution;
+using Aksio.IngressMiddleware.Security;
 
 namespace Aksio.IngressMiddleware;
